requirements.txt
python-docx
google-genai
flask
flask-cors
gunicorn
python-dotenv
